root/services/backend-api-service/requirements.txt
pydantic-settings
asyncpg
fastapi
requests
uvicorn
aiokafka
kafka-python
brotli
motor
umongo
